+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- New write-ups on Expo push notifications with Firebase and S3 images behind CloudFront.
+
 ## [0.3.0] (2026-09-15)
 
 ### Added